Updated ditaa to 0.11.0
Main new feature is SVG support!
This also involved changed the base image.  I unfortunately was not able
to get it to work correctly on an alpine openjdk image.  I wrestled for
a while with older versions and adding what seemed to be the missing
packages (fontconfig libfreetype6) but to no avail.